--- core/src/main/scala/codesearch/core/lexer/StringBuilder.scala
@@ -1,7 +1,0 @@
-package codesearch.core.lexer
-
-import codesearch.core.lexer.tokens.Token
-
-object StringBuilder {
-  def buildStringFromTokens(tokens: Seq[Token]) = tokens.map(_.repr).mkString
-}
