--- core/src/main/scala/codesearch/core/lexer/tokens/Content.scala
@@ -1,3 +1,0 @@
-package codesearch.core.lexer.tokens
-
-final case class Content(value: String) extends Token
