--- core/src/main/scala/codesearch/core/lexer/tokens/Other.scala
@@ -1,3 +1,0 @@
-package codesearch.core.lexer.tokens
-
-final case class Other(value: String) extends Token
