--- core/src/main/scala/codesearch/core/lexer/tokens/SpecialSymbol.scala
@@ -1,3 +1,0 @@
-package codesearch.core.lexer.tokens
-
-final case class SpecialSymbol(value: Char) extends Token
