+++ core/src/main/scala/codesearch/core/search/Search.scala
@@ -12,8 +12,8 @@ import cats.instances.list._
 import codesearch.core.index.directory.СSearchDirectory
 import codesearch.core.search.Search.{CSearchPage, CSearchResult, CodeSnippet, Package, PackageResult, snippetConfig}
 import codesearch.core.util.Helper.readFileAsync
-import codesearch.core.lexer.tokens.Token
-import codesearch.core.lexer.{StringBuilder, Tokenizer}
+import codesearch.core.index.regex.lexer.tokens.Token
+import codesearch.core.index.regex.lexer.{RowPicker, Tokenizer}
 
 import scala.sys.process.Process
 
@@ -60,7 +60,7 @@ trait Search {
 
   private def tokenizerAndBuilder(query: String): String = {
     val tokens: Seq[Token] = Tokenizer.parseStringWithSpecialSymbols(query)
-    StringBuilder.buildStringFromTokens(tokens)
+    RowPicker.buildStringFromTokens(tokens)
   }
 
   private def arguments(request: SearchRequest): List[String] = {
